techdebt2026: patchsets: add descriptive .md of students' patchsets
Maybe the most important part of the paper is when we bring the
interactions with the maintainers in regards to how they perceive the
proposed deduplications contributions. In this sense, add
`patchsets/notes.md` that focuses on describing the maintainers' views
and opinions on deduplications, specially the reasoning behind the
rejections.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>