techdebt2026: patchsets: anom-threads: add anonimized patchsets' threads
As we will be quoting maintainers to highlight some takeaways in terms
of why a deduplication would be rejected in favor of other factors, we
need to include the all the necessary interactions and contexts that
produced these quotes. Moreover, these need to be anonimized.

With that in mind, add the `patchsets/anom-threads` directory with the
anonimized threads of all versions of every patchset sent by the
students. These were manually anonimized by removing all the
email-related metadata (except the subjects), names, email adresses, and
dates.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>