techdebt2026: techdebt: add takeaways w/ example quotes
From the patchset analysis, we detected 4 takeaways beyond the one
already in place. For now, dump them along with quotes in a temporary
section "Takeaways (temporary)", just to gauge the size of adding those
boxes. Beyond adding the takeaways and quotes, fill the column
"Takeaway" from the table with the contributions.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>