techdebt2026: techdebt: change takeaway 4 to use correct reference
As pointed out in a TODO, the reference being used to base the claim
that jumping across code fragments increases cognitive load was
not the best one, and we needed to resolve the "base" reference the one
used cited. After some research, we settled with the Sillito et. al [1]
reference, as it better fits what we are trying to claim.

[1]: https://doi.org/10.1145/1181775.1181779

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>