techdebt2026: techdebt: rephrase paragraph about undetailed groups
Group 5 and Group 11 were not detailed in the same sense as the other
groups, as they did not produce or influence the takeaways and the core
idea of the paper. However, the way it was put, made it feel like they
had literally no meaning in this regard, which was not true, so rephrase
this paragraph for a better contextualization.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>