techdebt2026: patchsets: anom-threads: remove remaining PIIs
In the anonymized threads, there is still some Personally Identifiable
Information (PII), like some names, parts of URLs with
names/surnames/email adresses, and some sent times. Considering those,
remove them from the replication package.

Of course, if one is hellbent in finding the source of these threads,
this shouldn't be impossible, yet, we "what we control" is anonymized.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>